pyproject.toml
[project]
name = "gharanagold"
version = "1.0.0"
description = "GharanaGold ERP System"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.0",
    "flask-sqlalchemy==3.1.1",
    "flask-login==0.6.3",
    "flask-migrate==4.0.5",
    "flask-wtf==1.2.1",
    "email-validator==2.1.0.post1",
    "pymysql==1.1.0",
    "cryptography==42.0.5",
    "python-dotenv==1.0.1",
    "openpyxl==3.1.2",
    "resend==0.7.2",
    "psycopg2-binary==2.9.9",
    "supabase==2.3.4",
    "reportlab==4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
